+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Version 0.5.4 (TBA)
+* New Feature: Introduced trade callbacks
 
 # Version 0.5.3 (16.01.2018)
 * New Feature: Introduced the RawOrderBook Manager